pandas update

+++ cognite_synthetic_tags/tag_resolver.py
@@ -300,7 +300,7 @@ class TagResolver:
         # convert any non-series items to series:
         #   (repeat the item for every index)
         all_series: List[pd.Series] = [
-            val if isinstance(val, pd.Series) else pd.Series(val, index=index)
+            val if isinstance(val, pd.Series) else pd.Series(val, index=index, dtype=float)
             for val in data
         ]
